[fix]: menuId type 수정 (#64)

+++ app/src/main/java/com/kuit/ourmenu/data/model/home/response/HomeResponse.kt
@@ -14,7 +14,7 @@ data class HomeResponse(
 
 @Serializable
 data class RecommendMenuList(
-    val menuId: Int = 0,
+    val menuId: Long = 0,
     val menuTitle: String = "",
     val storeName: String = "",
     val menuImgUrl: String = "",
